// src/service/swagger-service/models/AuthResponse.ts
// tslint:disable
/**
 * Express.js Backend API
 * 一个专为前端工程师设计的Express.js后端API模板
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type {
    AuthResponseData,
} from './';

/**
 * @export
 * @interface AuthResponse
 */
export interface AuthResponse {
    /**
     * @type {boolean}
     * @memberof AuthResponse
     */
    success?: boolean;
    /**
     * @type {AuthResponseData}
     * @memberof AuthResponse
     */
    data?: AuthResponseData;
    /**
     * @type {string}
     * @memberof AuthResponse
     */
    message?: string;
}