// src/service/swagger-service/models/ApiAuthDeleteAccountDelete200Response.ts
// tslint:disable
/**
 * Express.js Backend API
 * 一个专为前端工程师设计的Express.js后端API模板
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface ApiAuthDeleteAccountDelete200Response
 */
export interface ApiAuthDeleteAccountDelete200Response {
    /**
     * @type {boolean}
     * @memberof ApiAuthDeleteAccountDelete200Response
     */
    success?: boolean;
    /**
     * @type {string}
     * @memberof ApiAuthDeleteAccountDelete200Response
     */
    message?: string;
}