// src/service/swagger-service/models/User.ts
// tslint:disable
/**
 * Express.js Backend API
 * 一个专为前端工程师设计的Express.js后端API模板
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface User
 */
export interface User {
    /**
     * 用户ID
     * @type {number}
     * @memberof User
     */
    id?: number;
    /**
     * 用户姓名
     * @type {string}
     * @memberof User
     */
    name: string;
    /**
     * 用户邮箱
     * @type {string}
     * @memberof User
     */
    email: string;
    /**
     * 用户年龄
     * @type {number}
     * @memberof User
     */
    age?: number;
    /**
     * 用户头像URL
     * @type {string}
     * @memberof User
     */
    avatar?: string;
}